+++ NPM Release/startsWithCapital.js
@@ -1,5 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.startsWithCapital = void 0;
-function startsWithCapital(word) { return word.charAt(0) === word.charAt(0).toUpperCase(); }
-exports.startsWithCapital = startsWithCapital;