+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version `1.3.0`
+- Special characters are now used to make the code less readable
+- New string encoding used in some cases
+- Another string encoding used in another cases
+> NOTE: The old string encoding remains to be used (in some cases)
+- The config now has an `epicEndArt` property
+- Inform function improved
+- All scripts moved to the script/ directory
+- Improved README.md
+- NPM Release folder is no longer on github
+
 # Version `1.2.1`
 - Fixed readme
 - Help will now be printed if you have an error with arguments